+++ src/main/java/net/sacredlabyrinth/Phaed/PreciousStones/managers/UnbreakableManager.java
@@ -370,7 +370,7 @@ public final class UnbreakableManager
                     {
                         revertedCount++;
                         block.setTypeId(unbreakable.getTypeId());
-                        block.setData(unbreakable.getData());
+                        block.setData((byte)unbreakable.getData());
                     }
                 }
             }
